refactored the last 1/3
